+++ src/libsodium/crypto_box/curve25519xsalsa20poly1305/ref/after_curve25519xsalsa20poly1305.c
@@ -3,7 +3,7 @@
 
 int crypto_box_afternm(
   unsigned char *c,
-  const unsigned char *m,unsigned long long mlen,
+  const unsigned char *m,uint64_t mlen,
   const unsigned char *n,
   const unsigned char *k
 )
@@ -13,7 +13,7 @@ int crypto_box_afternm(
 
 int crypto_box_open_afternm(
   unsigned char *m,
-  const unsigned char *c,unsigned long long clen,
+  const unsigned char *c,uint64_t clen,
   const unsigned char *n,
   const unsigned char *k
 )
