+++ src/libsodium/crypto_secretbox/xsalsa20poly1305/ref/box_xsalsa20poly1305.c
@@ -4,7 +4,7 @@
 
 int crypto_secretbox(
   unsigned char *c,
-  const unsigned char *m,unsigned long long mlen,
+  const unsigned char *m,uint64_t mlen,
   const unsigned char *n,
   const unsigned char *k
 )
@@ -19,7 +19,7 @@ int crypto_secretbox(
 
 int crypto_secretbox_open(
   unsigned char *m,
-  const unsigned char *c,unsigned long long clen,
+  const unsigned char *c,uint64_t clen,
   const unsigned char *n,
   const unsigned char *k
 )
