+++ src/libsodium/crypto_stream/aes128ctr/portable/int128.h
@@ -4,8 +4,8 @@
 #include "common.h"
 
 typedef struct{
-  unsigned long long a;
-  unsigned long long b;
+  uint64_t a;
+  uint64_t b;
 } int128;
 
 #define xor2 crypto_stream_aes128ctr_portable_xor2
