+++ src/libsodium/include/sodium/crypto_hash_sha256.h
@@ -2,6 +2,7 @@
 #define crypto_hash_sha256_H
 
 #include <stddef.h>
+#include <stdint.h>
 #include "export.h"
 
 #define crypto_hash_sha256_BYTES 32U
@@ -18,7 +19,7 @@ SODIUM_EXPORT
 const char * crypto_hash_sha256_primitive(void);
 
 SODIUM_EXPORT
-int crypto_hash_sha256(unsigned char *,const unsigned char *,unsigned long long);
+int crypto_hash_sha256(unsigned char *,const unsigned char *,uint64_t);
 
 #ifdef __cplusplus
 }
