+++ src/libsodium/include/sodium/crypto_hashblocks_sha512.h
@@ -2,6 +2,7 @@
 #define crypto_hashblocks_sha512_H
 
 #include <stddef.h>
+#include <stdint.h>
 #include "export.h"
 
 #define crypto_hashblocks_sha512_STATEBYTES 64U
@@ -21,7 +22,7 @@ SODIUM_EXPORT
 const char * crypto_hashblocks_sha512_primitive(void);
 
 SODIUM_EXPORT
-int crypto_hashblocks_sha512(unsigned char *,const unsigned char *,unsigned long long);
+int crypto_hashblocks_sha512(unsigned char *,const unsigned char *,uint64_t);
 
 #ifdef __cplusplus
 }
