+++ src/libsodium/include/sodium/crypto_shorthash_siphash24.h
@@ -2,6 +2,7 @@
 #define crypto_shorthash_siphash24_H
 
 #include <stddef.h>
+#include <stdint.h>
 #include "export.h"
 
 #define crypto_shorthash_siphash24_BYTES 8U
@@ -18,7 +19,7 @@ SODIUM_EXPORT
 const char * crypto_shorthash_siphash24_primitive(void);
 
 SODIUM_EXPORT
-int crypto_shorthash_siphash24(unsigned char *,const unsigned char *,unsigned long long,const unsigned char *);
+int crypto_shorthash_siphash24(unsigned char *,const unsigned char *,uint64_t,const unsigned char *);
 
 #ifdef __cplusplus
 }
