+++ src/libsodium/include/sodium/crypto_sign_edwards25519sha512batch.h
@@ -2,6 +2,7 @@
 #define crypto_sign_edwards25519sha512batch_H
 
 #include <stddef.h>
+#include <stdint.h>
 #include "export.h"
 
 #define crypto_sign_edwards25519sha512batch_SECRETKEYBYTES 64U
@@ -25,10 +26,10 @@ SODIUM_EXPORT
 const char * crypto_sign_edwards25519sha512batch_primitive(void);
 
 SODIUM_EXPORT
-int crypto_sign_edwards25519sha512batch(unsigned char *,unsigned long long *,const unsigned char *,unsigned long long,const unsigned char *);
+int crypto_sign_edwards25519sha512batch(unsigned char *,uint64_t *,const unsigned char *,uint64_t,const unsigned char *);
 
 SODIUM_EXPORT
-int crypto_sign_edwards25519sha512batch_open(unsigned char *,unsigned long long *,const unsigned char *,unsigned long long,const unsigned char *);
+int crypto_sign_edwards25519sha512batch_open(unsigned char *,uint64_t *,const unsigned char *,uint64_t,const unsigned char *);
 
 SODIUM_EXPORT
 int crypto_sign_edwards25519sha512batch_keypair(unsigned char *,unsigned char *);
